// src/gen/predictionmarket/poa/v1/tx.ts
export interface ValidatorSlot {
  index: number;
  name: string;
  consensusAddress: string;
  consensusPubKey: Uint8Array;
  power: bigint;
}

export interface MsgSetValidatorSet {
  authority: string;
  validators: ValidatorSlot[];
}

export interface MsgSetValidatorSetResponse {
}

export interface MsgUpdateAdmin {
  authority: string;
  newAdmin: string;
}

export interface MsgUpdateAdminResponse {
}

function createBaseValidatorSlot(): ValidatorSlot {
  return { index: 0, name: "", consensusAddress: "", consensusPubKey: new Uint8Array(0), power: 0n };
}

export const ValidatorSlot: MessageFns<ValidatorSlot> = {
  encode(message: ValidatorSlot, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.index !== 0) {
      writer.uint32(8).uint32(message.index);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.consensusAddress !== "") {
      writer.uint32(26).string(message.consensusAddress);
    }
    if (message.consensusPubKey.length !== 0) {
      writer.uint32(34).bytes(message.consensusPubKey);
    }
    if (message.power !== 0n) {
      if (BigInt.asIntN(64, message.power) !== message.power) {
        throw new globalThis.Error("value provided for field message.power of type int64 too large");
      }
      writer.uint32(40).int64(message.power);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ValidatorSlot {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseValidatorSlot();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.index = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.consensusAddress = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.consensusPubKey = reader.bytes();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.power = reader.int64() as bigint;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<ValidatorSlot>, I>>(base?: I): ValidatorSlot {
    return ValidatorSlot.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ValidatorSlot>, I>>(object: I): ValidatorSlot {
    const message = createBaseValidatorSlot();
    message.index = object.index ?? 0;
    message.name = object.name ?? "";
    message.consensusAddress = object.consensusAddress ?? "";
    message.consensusPubKey = object.consensusPubKey ?? new Uint8Array(0);
    message.power = object.power ?? 0n;
    return message;
  },
};

function createBaseMsgSetValidatorSet(): MsgSetValidatorSet {
  return { authority: "", validators: [] };
}

export const MsgSetValidatorSet: MessageFns<MsgSetValidatorSet> = {
  encode(message: MsgSetValidatorSet, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.authority !== "") {
      writer.uint32(10).string(message.authority);
    }
    for (const v of message.validators) {
      ValidatorSlot.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgSetValidatorSet {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgSetValidatorSet();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.authority = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.validators.push(ValidatorSlot.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<MsgSetValidatorSet>, I>>(base?: I): MsgSetValidatorSet {
    return MsgSetValidatorSet.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgSetValidatorSet>, I>>(object: I): MsgSetValidatorSet {
    const message = createBaseMsgSetValidatorSet();
    message.authority = object.authority ?? "";
    message.validators = object.validators?.map((e) => ValidatorSlot.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const MsgUpdateAdmin: MessageFns<MsgUpdateAdmin> = {
  encode(message: MsgUpdateAdmin, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.authority !== "") {
      writer.uint32(10).string(message.authority);
    }
    if (message.newAdmin !== "") {
      writer.uint32(18).string(message.newAdmin);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgUpdateAdmin {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgUpdateAdmin();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.authority = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.newAdmin = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<MsgUpdateAdmin>, I>>(base?: I): MsgUpdateAdmin {
    return MsgUpdateAdmin.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgUpdateAdmin>, I>>(object: I): MsgUpdateAdmin {
    const message = createBaseMsgUpdateAdmin();
    message.authority = object.authority ?? "";
    message.newAdmin = object.newAdmin ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}